Use u64 for byte ranges/array index/shape etc.

+++ examples/array_write_read.rs
@@ -55,9 +55,12 @@ fn array_write_read() -> Result<(), Box<dyn std::error::Error>> {
         .into_par_iter()
         .map(|i| {
             let chunk_grid: &Box<dyn ChunkGridTraits> = array.chunk_grid();
-            let chunk_indices: Vec<usize> = vec![i, 0];
+            let chunk_indices: Vec<u64> = vec![i, 0];
             let chunk_subset: ArraySubset = chunk_grid.subset(&chunk_indices, &array.shape())?;
-            array.store_chunk_elements(&chunk_indices, &vec![i as f32; chunk_subset.num_elements()])
+            array.store_chunk_elements(
+                &chunk_indices,
+                &vec![i as f32; chunk_subset.num_elements() as usize],
+            )
             // let chunk_shape = chunk_grid.chunk_shape(&chunk_indices, &array.shape())?;
             // let chunk_array = ndarray::ArrayD::<f32>::from_elem(chunk_shape.clone(), i as f32);
             // array.store_chunk_ndarray(&chunk_indices, &chunk_array.view())
